--- cfron-web/.gitignore
@@ -1,18 +1,0 @@
-/target
-**/*.rs.bk
-Cargo.lock
-bin/
-wasm-pack.log
-out/
-node_modules/
-.yarn/
-.yarnrc
-
-/pkg/*
-!/pkg/cfron.d.ts
-!/pkg/cfron.js
-!/pkg/LICENSE
-!/pkg/package.json
-!/pkg/README.md
-
-/tests/bundle
