--- cfron-web/package.json
@@ -1,18 +1,0 @@
-{
-  "private": true,
-  "engines": {
-    "node": ">=14.0.0",
-    "yarn": ">=1.22.0"
-  },
-  "scripts": {
-    "wasm-build": "node ./scripts/wasm-build.mjs",
-    "pretest": "yarn wasm-build && yarn install --force && webpack",
-    "test": "jest"
-  },
-  "devDependencies": {
-    "@cloudflare/cfron": "file:./pkg",
-    "jest": "^26.6.3",
-    "webpack": "^5.9.0",
-    "webpack-cli": "^4.2.0"
-  }
-}
