--- cfron-worker/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-authors = ["Aaron Loyd <aloyd@cloudflare.com>"]
-description = """
-A fallback Worker for validating, describing, and
-getting the next execution time of cron strings.
-"""
-edition = "2018"
-license-file = "LICENSE"
-name = "cfron-worker"
-repository = "https://github.com/cloudflare/cfron"
-version = "0.1.0"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[dependencies]
-cfron = {path = "../cfron", version = "0.1"}
-chrono = {version = "0.4", features = ["wasmbind"]}
-console_error_panic_hook = {version = "0.1"}
-js-sys = "0.3"
-wasm-bindgen = {version = "=0.2.65"}
-
-[dev-dependencies]
-wasm-bindgen-test = "0.2"
