--- cfron/fuzz/Cargo.toml
@@ -1,26 +1,0 @@
-
-[package]
-name = "cfron-fuzz"
-version = "0.0.0"
-authors = ["Automatically generated"]
-publish = false
-edition = "2018"
-
-[package.metadata]
-cargo-fuzz = true
-
-[dependencies]
-libfuzzer-sys = "0.3"
-
-[dependencies.cfron]
-path = ".."
-
-# Prevent this from interfering with workspaces
-[workspace]
-members = ["."]
-
-[[bin]]
-name = "default"
-path = "fuzz_targets/default.rs"
-test = false
-doc = false
