--- cfron/fuzz/fuzz_targets/default.rs
@@ -1,9 +1,0 @@
-#![no_main]
-use libfuzzer_sys::fuzz_target;
-use cfron::parse::CronExpr;
-
-fuzz_target!(|data: &[u8]| {
-    if let Ok(s) = std::str::from_utf8(data) {
-        let _ = s.parse::<CronExpr>();
-    }
-});
