+++ saffron-web/.gitignore
@@ -1,0 +1,18 @@
+/target
+**/*.rs.bk
+Cargo.lock
+bin/
+wasm-pack.log
+out/
+node_modules/
+.yarn/
+.yarnrc
+
+/pkg/*
+!/pkg/saffron.d.ts
+!/pkg/saffron.js
+!/pkg/LICENSE
+!/pkg/package.json
+!/pkg/README.md
+
+/tests/bundle
