+++ saffron/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+authors = ["Aaron Loyd <aloyd@cloudflare.com>"]
+description = "A Quartz-like cron parser used as part of Cron Triggers in Cloudflare Workers"
+edition = "2018"
+license-file = "LICENSE"
+name = "saffron"
+repository = "https://github.com/cloudflare/saffron"
+version = "0.1.0"
+
+[features]
+default = []
+std = []
+
+[[bench]]
+harness = false
+name = "cron"
+
+[[example]]
+name = "future-times"
+required-features = ["chrono/clock"]
+
+[dependencies]
+chrono = {version = "0.4", default-features = false}
+nom = {version = "5.1", default-features = false}
+
+[dev-dependencies]
+criterion = "0.3"
