+++ saffron/fuzz/Cargo.toml
@@ -1,0 +1,26 @@
+
+[package]
+name = "saffron-fuzz"
+version = "0.0.0"
+authors = ["Automatically generated"]
+publish = false
+edition = "2018"
+
+[package.metadata]
+cargo-fuzz = true
+
+[dependencies]
+libfuzzer-sys = "0.3"
+
+[dependencies.saffron]
+path = ".."
+
+# Prevent this from interfering with workspaces
+[workspace]
+members = ["."]
+
+[[bin]]
+name = "default"
+path = "fuzz_targets/default.rs"
+test = false
+doc = false
